Hacky solution to make categorization work
Basically, when the Category of a Meme was changed, the MemeService's instance of the DbContext would discover the Category. However, the Category is already being tracked by the CategoryService's DbContext instance. To prevent this, I'm having them share the same DbContext instance which is absolutely NOT how EF wants to be used (per https://docs.microsoft.com/en-us/ef/core/dbcontext-configuration/#the-dbcontext-lifetime) but it works so what the hell.

Known issues:
- The UI doesn't update to reflect DB changes. Until I figure out how to do this the right way, I'm gonna make a notifier that'll have the relevant classes re-run their search functions
- You can only click the leaf nodes of the context menu. Any menu items with children won't fire their command.