Changing tags is now possible!!!!!!!!!
Holy shit this made me want to just give up. I worked on this on and off for months just trying to get dialogs to work like they should. It really shouldn't be this hard. After that, it was just a matter of creating a working tagging service and hooking it up to the UI.