Fixed a bug with drag & drop categorization
Fixed a bug where a meme couldn't be dragged and dropped onto any kind of subcategory. It seems that, in my infinite wisdom, I thought it would be a good idea to just chain three GetLogicalParent() calls together. It was predicated on the logical ancestry being a certain length and would OBVIOUSLY fail for a subcategory.

Clearly my past self was juuuuust smart enough to realize something might be wrong with this. Smart enough to leave a comment but not smart enough to actually double-check whether this would work.

Come to think of it, I'm almost certain this used to work because there's no way I'd stop working on the categories feature to add tagging if categorization was still broken. Now I'm worried that there's something else that's way more broken because of the sheer number of changes I've made .-.