Limit the number of memes shown on the "home" screen
When no category is selected and no search terms are entered, the memes displayed are the most recently imported memes. If the library is huge, EF will throw an error on startup. This is because the categories are having their navigation properties accessed to draw the categories view while the memes list is still being populated. This will eventually be solved with #69. However, I think there may also be a slight performance benefit to doing this.