Attempted a POC with realm
It didn't go too well and I basically have to gut the entire application and use CommunityToolkit instead of ReactiveUI. I'm too frustrated right now to finish doing that and frankly, I hate developing with AXAML because it's such a pain in the ass to figure out (likely due to my own ignorance). So, I'm going to put this on its own branch in case I fail in my next endeavor. What might that endeavor be? I'm going to rewrite the entire project in Flutter because I hate myself so much! :D