Implement Windows Explorer-like selection logic
I was disappointed that no mechanism like this is already available in Flutter. However, it wasn't TOO awful to implement manually. I also found a really impressive flaw with Flutter where, if onDoubleTap is defined on an InkWell, the onTap function gets called with an added 300ms of latency to allow for detection of double taps.